signature vectors are short not double

+++ ADOL-C/include/adolc/interfaces.h
@@ -435,12 +435,12 @@ ADOLC_DLL_EXPORT int int_reverse_safe
 /*--------------------------------------------------------------------------*/
 ADOLC_DLL_EXPORT int get_num_switches(short);
 ADOLC_DLL_EXPORT int zos_an_forward(short,int,int,int,const double*,double*,double*);
-ADOLC_DLL_EXPORT double firstsign(int, double, double*);
-ADOLC_DLL_EXPORT double ext_firstsign(double,double,int,double,double*);
+ADOLC_DLL_EXPORT short firstsign(int, double, double*);
+ADOLC_DLL_EXPORT short ext_firstsign(double,double,int,double,double*);
 ADOLC_DLL_EXPORT int fos_an_forward(short,int,int,const double*,double*,double*,double*,double*,double*);
 ADOLC_DLL_EXPORT int fov_an_forward(short,int,int,int,const double*,double**,double*,double**,double*,double**);
-ADOLC_DLL_EXPORT int fos_an_sig_forward(short,int,int,const double*,double*,int,double*,double*,double*,double*,double*,double*,double*);
-ADOLC_DLL_EXPORT int fov_an_sig_forward(short,int,int,int,const double*,double**,int,double*,double*,double*,double**,double*,double**,double*);
+ADOLC_DLL_EXPORT int fos_an_sig_forward(short,int,int,const double*,double*,int,short*,short*,double*,double*,double*,double*,short*);
+ADOLC_DLL_EXPORT int fov_an_sig_forward(short,int,int,int,const double*,double**,int,short*,short*,double*,double**,double*,double**,short*);
 /*--------------------------------------------------------------------------*/
 ADOLC_DLL_EXPORT int fos_an_reverse(short,int,int,int,int,double*);
